+++ R/get_signalp.R
@@ -16,7 +16,7 @@
 #' @param splitter An integer indicating the number of sequences to be in each .fasta file that is to be sent to the server. Defaults to 500. Change only in case of a server side error. Accepted values are in range of 1 to 2000.
 #' @param sleep A numeric indicating the pause in seconds between POST and GET server calls, at default set to 1s. Decreasing is not recommended.
 #' @param attempts Integer, number of attempts if server unresponsive, at default set to 2.
-#' @param progress Bolean, whether to show the progress bar, at default set to FALSE.
+#' @param progress Boolean, whether to show the progress bar, at default set to FALSE.
 #' @param ... currently no additional arguments are accepted apart the ones documented bellow.
 #'
 #' @return  A data frame with columns:
@@ -53,7 +53,7 @@
 #' @import seqinr
 #' @import httr
 #' @import xml2
-#' @export get_signalp
+#' @export 
 
 get_signalp <- function (data, ...){
   if (missing(data) || is.null(data)) get_signalp.default(...)